Ajout d'une jauge de vie

+++ src/views/monsters/CircleMonsterView.java
@@ -38,7 +38,7 @@ public class CircleMonsterView extends MonsterView {
 		GL11.glDisable(GL11.GL_TEXTURE_2D);
 		GL11.glColor3ub((byte)actualColor.getRed(),(byte)actualColor.getGreen(),(byte)actualColor.getBlue());
 		GL11.glTranslated(getX(), getY(), 0);
-		cylinder.draw((float)(size/2),0,height, 50, 1);
+		cylinder.draw((float)(size/2),0,(float)height, 50, 1);
 		GL11.glTranslated(-getX(), -getY(), 0);
 		paintChildren();
